--- src/main/java/team/silvertown/masil/mate/repository/MateRepository.java
@@ -1,8 +1,0 @@
-package team.silvertown.masil.mate.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import team.silvertown.masil.mate.domain.Mate;
-
-public interface MateRepository extends JpaRepository<Mate, Long> {
-
-}
