+++ src/main/java/team/silvertown/masil/mate/repository/participant/MateParticipantRepository.java
@@ -1,4 +1,4 @@
-package team.silvertown.masil.mate.repository;
+package team.silvertown.masil.mate.repository.participant;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.silvertown.masil.mate.domain.MateParticipant;
